requirements.txt
discord.py[voice]==2.7.1
yt-dlp>=2024.1.1
python-dotenv==1.0.0
